move everything to graphql types
