+++ src/components/Item/index.tsx
@@ -1,7 +1,7 @@
 import React from 'react';
 import styled from 'styled-components';
 import formatDistanceToNowStrict from 'date-fns/formatDistanceToNowStrict';
-import { PantryItem, Quantity } from '../../schema/pantry';
+import { PantryItem, Quantity } from '../../schema/generated';
 import { Freshness } from '../Freshness';
 
 const ItemContainer = styled.div<{ empty?: boolean }>`
